Check if the input from the user matches the regex
We check if the regex was matched by what the user input.
If it doesn't match, a message is printed which informs the user.

The branch for when the regex is matched is not yet implemented,
but the idea is that it will then pass that input to be properly parsed
into game data with which the board can be populated.