Loop section for user's input
If the user's input does not match the regex,
then they should be prompted to enter it again.
Thus the section prompting the user for input and checking it against
the regex is placed inside an infinite loop.

With further design, the loop will be broken out of when the input
matches the regex.