Change the regex so that the ending semicolon is optional
Rather than requiring the user to enter a semicolon for each and every
token, I think it makes more sense to only require one between each
token. I have made the ending semicolon optional.

I could have changed the regex which matches the ending token so
that the user could enter any character sequence after the co-ordinates
without it causing the pattern to fail.

The benefit of doing that would be that the pattern wouldn't depend on
what's entered after the actuall important content which we're trying
to match; namely the co-ordinates.

An arguable downside is that the pattern shouldn't accept anything
other than the tokens necessary for the user's ships.
This is largely a philisophical argument over which is the better
choice, but I've chosen to go with only allowing the tokens which
matter and whitespace to be accepted by the match.