Name the first group matching the carrier in the regex
I have named the group which matches the carrier's co-ordinates in
the regex applied to the input string.

I will then go through the other ships, naming the groups which match
each ship. This will then allow the matched parts of the string to be
seperated easily and then used to build the game's data model