Display the board during initialising the GameBoard class
When initialising the GameBoard class, the board should be displayed.

At the moment the function to display the board is not implemented,
but the idea is that it will show the board and then ask the player
for the co-ordinates of the each of their ships, displaying the board
after each ships co-ordinates are defined.