Remove prompts asking for ship co-ordinates from main function
I've decided to move the starting prompts for the user to input
their ship's positions on the board. I'm going to move it in to the
Game class.

The idea is then to have the Game class construct the boards for
both the player and the cpu and for it to contain the logic for
the rest of the game.