Check if the input co-ordinates for the carrier can be parsed
The setting up of the player's ships branch on whether the input
string from the user could be matched against a regex which matches
the format of the co-ordinates, and groups the values of interest,
so that they can be retreived.

The call to check the input then returns either `None` for a failed
match, or a dictionary containing the parsed tokens for the
co-ordinates entered.