Add a private method to PlayerGameBoard to normalise the co-ords to one form
My regex allows for the user to input the co-ordinates capitalised
however they wish to. They can also choose to either ender the full
word for the direction of their ship from the starting co-ordinate,
or just enter the first letter.

This allows for flexibility of input, however that then causes the
potential inputs to be a much more difficult set to work with when
converting the string tokens into numbers, or some other data.

To reduce this complexity, I've implemented this 'normalisation'
function, to convert all the different variations of input that
one co-ordinate can have into one "canonical" version.