Add a status field to the dictionary repressenting the ship
The status field repressents the status of each square that the ship
occupies.

It is a list of booleans, where True means that that part of the ship
is still healthy.

This field will be updated when one of the ship's squares is fired on,
and will be read from when drawing the board, and when checking if the
ship as a whole is still afloat or has been sunk.