Add private method to PlayerGameBoard to check that a ship does not overlap with previous ships
Before I can accept a users choice of placement for their ship,
I need to make sure that the input is valid as co-ordinates, that the
placement of the ship would actually fit on the board, and that there
is nothing already on the squares which the new board is being
considered for.

This function answers this third requirement, taking a ship and checking
that there is nothing already on the squares the ship would otherwise
go on.