Add a ship to the board once all the placement checks have passed
Once the input from the user has passed through the checks to
make sure that the ship specified can be added to the squares
which were requested, we add the ship to the player board.

The infinite loop which all of this code is inside is there to ask
the user to enter the co-ordinates in the cases where the input did
not pass the checks.
Now that we've got this far, we can break out of the loop and move on
to the next ship.