Correct syntax error when updating the coords entry in ship_dict in PlayerGameBoard.__normalise_coords()
I managed to commit a change which introduces a syntax error
in to the source.

dict.update() expects a dictionary as it's input, which it then
unions with dict.
But, I put the colon and brace in the wrong place, so it didn't form a
dictionary.

This change resolves this.