Move the section to ask the user for the ships coords into a private method in PlayerGameBoard
Move the section which prompts the user for the co-ordinates for the
carrier and runs their input through the checkers.
It is moved into a private method in the PlayerGameBoard class, and
the mention of the carrier and the ship size are abstracted to
parameters.

This means that I can now run the same method on all ship classes
and sizes. LESS REPITITION!