Change the representation of elements of __board to be tuples of strings and dictionaries
There was a bug in my code, were the print_shipboard method could not
join the elements which I had changed to tuples by "|"s.

The join method was expecting to join on strings, but obviously,
it can't do that when I pass it a tuple.

To fix this, instead of passing the join function the elements of
the list which I'm iterating over, I pass it a list of the first
member of the tuples in the list. This happens to be a string (char),
so they can be joined together.

Of course, to do that without having to do some sort of
pattern matching, or other fiddly stuff, I would have to make all
the elements of the list into tuples of this form.

I changed the initialisation of the elements of the list by
assigning them the following default tuple:
(" ", None)

To do this without making another bug, I have to change the code
which checks for if the squares of the board are empty, in order to
avoid ships overlapping.
Now I need to also take the first element out of the tuple to
compare it to " ".