Correct a bug with the direction of the vertical axis being reversed
I made a classic mistake of misinterpreting the indexing of a list
against the indexing used in a co-ordinate system.

What I mean by this is that the co-ordinate system which I settled on
is for the top left corner of the board to be a1 (internally this
is represented as (0,0)). However, I had been thinking of the up
direction of a ship meaning that the numbers would increase as you
go along the squares covered by the ship.

But, that means that going higher in number in the vertical direction
looks like the ship is going down on the screen!
To remedy this, I changed the direction that I defined up and down
to be, essentially flipping them. Up is lower in number, and down is
higher in number.

I could have decided to have (0,0) a1 be the bottom left corner.
You might think that this is the most intuitive choice.
However, while by no means impossible, it is a little bit fiddly
to get this to work, as the function which prints the board, does
so from the start of the array to the end. You could reverse this,
which would require using a formular like: '10 - i'