Fix a bug where the ships are mirror on the major diagonal
I found that the ships which I was declaring were being flipped on the
major diagonal, so that d/dx = -y

I managed this through another basic mistake of thinking of the
vertical axis as the outer lists in the board list, which doesn't
match up with how the lists are printed to the terminal.
The terminal will print in horizontal lines, of course, so I should
think of the outer lists as the horizontal axis, and so a change in
which outer list you're focusing on, is a change in the vertical axis.

Code can be confusing sometimes :P