Add summary of the 'Entry of unit positions which overlap with previously defined units' to 'README.md'
This section clarifies the programs handling of the error case were
a user tries to define the position of one of their units, such that
the unit would overlap with a previously defined unit.

It also explains the intended value to the user of the way which the
program handles this error.